Move FrontendOptions and FrontendInputs routines to FrontendOptions.cpp
